HACK: android: binder: remove stats accounting and tracing
We already removed debugfs / dmesg interfaces to remove transaction
logging, so it's useless to keep track of each transaction when you
can't access it.

As for tracing, even if the config includes ftrace, I don't think it's
needed anyway.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>

[kdrag0n: update to latest android-4.4 binder, clean up traces of stats.
I'll probably bring tracing back because it's zero-overhead when FTRACE
is off.]
Signed-off-by: kdrag0n <dragon@khronodragon.com>