Merge branch 'spf'
This adds Speculative Page Faults from upstream linux-mm patchsets (not
merged into mainline at this time) backported to 4.9 by CAF. These
commits are available in CAF's sdm845 branch as of the following tag:
LA.UM.7.3.r1-06300-sdm845.0

SPF should offer up to a 15% improvement in app opening speeds as
described by CAF developer Ganesh Mahendran:
"We ported the SPF to kernel 4.9 in android devices.
For the app launch time, It improves about 15% average. For the apps
which have hundreds of threads, it will be about 20%.

Thanks."