Merge branch 'tcp-jiffies32'
This merge adds usage of jiffies32 in the TCP stack where appropriate
instead of tcp_time_stamp.

I skipped the full switch over to an internal 1 us precision TCP clock
(which is commit 9a568de4818d "tcp: switch TCP TS option (RFC 7323) to
1ms clock") because the clock is mostly exposed with 1 ms precision. Our
current system-wide clock is 1000 Hz, or 1 ms per tick, so this is not
necessary. It is extra work, delta from linux-stable, and CPU usage to
backport the full microsecond clock system just for TCP. I may
potentially revisit this in the future if we diverge from a 1000 Hz
system clock for whatever reason.