Makefile: Use -O2 with GCC-LTO
-O3 have aggressive optimizations that are considers to be only safe if certain predictions about register use, how stack frames are interacted with, and function reentrancy are true or not.

These predictions are not assured to be true in some code like the kernel especially when inline is used and usually it is mainly used for certain applications than kernel.

It's a really bad idea for an arm/arm64 platform to be using -O3... it is only meant for server side usage... arm is not a server machine after all.. or else all OEMs (including Google themselves) will beginning to use -O3... which never happened... that's my point of view.. convince me.. or deal with it.

-O2 : Optimize as much as possible, without taking the risk of significantly increasing the binary size or degrading performance. That's perfectly fine for me.