Defconfig: Disable PEQ by default
Since now most of the drivers are using system_power_efficient_wq function so.. we want to have stock behaviors of those drivers which will allow any CPU to get used whenever one of those drivers gets used.. of course user can toggle them on or off through userspace.

Workqueues can be performance or power-oriented. Currently, most workqueues are
bound to the CPU they were created on. This gives good performance (due to cache
effects) at the cost of potentially waking up otherwise idle cores (Idle from
scheduler's perspective. Which may or may not be physically idle) just to
process some work. To save power, we can allow the work to be rescheduled on a
core that is already awake.

Workqueues created with the WQ_UNBOUND flag will allow some power savings.
However, we don't change the default behavior of the system. so keep we keep it disabled to maintain stock behavior of those drivers/functions.