simple_lmk: Consider all positive adjs when finding victims
We are allowed to kill any process with a positive adj, so we shouldn't
exclude any processes with adjs greater than 999. This would present a
problem with quirky applications that set their own adj score, such as
stress-ng. In the case of stress-ng, it would set its adj score to 1000
and thus exempt itself from being killed by Simple LMK. This shouldn't
be allowed; any process with a positive adj, up to the highest positive
adj possible (32767) should be killable.

@THEBOSS619: Adapted it due to this commit [https://github.com/THEBOSS619/Note9-Zeus-Q10.0/commit/851775ace4a9fad402b524898ee158000e806990] is used for SLMK monitoring through sysfs

Reported-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>